move hooks into correct directory
