+++ src/pages/workspaces/workspace/Dashboard/AzureStorageDetails.ts
@@ -5,8 +5,8 @@ import { TooltipCell } from 'src/components/table';
 import { ReactComponent as AzureLogo } from 'src/images/azure.svg';
 import { getRegionFlag, getRegionLabel } from 'src/libs/azure-utils';
 import { AzureContext } from 'src/libs/workspace-utils';
+import { StorageDetails } from 'src/pages/workspaces/hooks/useWorkspace';
 import { InfoRow } from 'src/pages/workspaces/workspace/Dashboard/InfoRow';
-import { StorageDetails } from 'src/pages/workspaces/workspace/useWorkspace';
 
 interface AzureStorageDetailsProps {
   azureContext: AzureContext;
